% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_intensity.R
\name{choose_measure}
\alias{choose_measure}
\title{Choose Measurement, Region and Frequency for Intensity Reading}
\usage{
choose_measure(region = c("Entire Cell", "Membrane", "Nucleus"),
  freq = c("520", "540", "570", "620", "650", "690"),
  measure = c("Mean", "Min", "Max", "Std Dev", "Total"))
}
\arguments{
\item{region}{Region of interest, one of: "Entire Cell", "Membrane",
"Nucleus"}

\item{freq}{Frequency of interest, integer or character, one of: 520, 540,
570, 620, 650, 690.}

\item{measure}{Measurement of interest, one of: "Mean", "Min", "Max",
"Std Dev", "Total".}
}
\description{
Helper function for selecting region, frequency, and measure for cell
intensity readings according to inForm column name output.
}
