% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_intensity.R
\name{color_measure}
\alias{color_measure}
\title{Color Palette by Wavelength}
\usage{
color_measure(wl = NULL)
}
\arguments{
\item{wl}{Wavelength(s) or column name(s) containing wavelength.}
}
\description{
Uses colors chosen from \url{https://www.nature.com/articles/nmeth.1618} as
suggested in \url{https://petebankhead.gitbooks.io/imagej-intro/content/chapters/colors/colors.html}.
}
\examples{
color_measure()
color_measure("Entire Cell Opal 540 Mean")
color_measure(540)

}
