% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_observations}
\alias{count_observations}
\alias{count_cells}
\title{Count Observations}
\usage{
count_observations(x, count_by = Phenotype, expected = NULL,
  exclude = NULL, include_totals = FALSE)

count_cells(csd, count_by = Phenotype, expected = NULL,
  exclude = NULL, include_totals = FALSE)
}
\arguments{
\item{x}{Input data frame}

\item{count_by}{Column to count by, unquoted}

\item{expected}{Expected values for the \code{count_by} columns. Providing these
values ensures that 0 counts are included even if a value never appears
in the input data.}

\item{exclude}{Excludes values in the \code{count_by} column entirely.}

\item{include_totals}{Adds a "{count_by} Total" value to the returned data.}
}
\value{
An ungrouped data frame with counts by grouping variables in addition
to \code{count_by} column. Zero counts are included.
}
\description{
Counts observations by \code{count_by} column, respecting grouping variables.
When the input data frame is cell segmentation data and each row is an
individual cell, then the output is cell counts.
}
\section{Functions}{
\itemize{
\item \code{count_cells}: Counts cells using \code{count_observations()} and
returns a data frame with a \code{Cell Count} column in place of \code{n}.
}}

\examples{
\dontrun{
csd \%>\% group_by(Sample) \%>\% count_observations(Phenotype)
csd \%>\% group_by(Sample, `Tissue Category`) \%>\% count_cells(Phenotype)
}

}
