% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_intensity.R
\name{gather_cell_intensity}
\alias{gather_cell_intensity}
\title{Gathers Cell Intensity Readings}
\usage{
gather_cell_intensity(csd, ..., exclude = "Undefined",
  summarize = length(group_vars(csd)))
}
\arguments{
\item{csd}{Cell Segmentation Data}

\item{...}{Named variable pairs, e.g. \code{CD3 = choose_measure("Membrane", 520, "Mean")}
or \code{CD3 = "Membrane Opal 520 Mean"}}

\item{exclude}{Exclude intensity readings from cells with these \code{Phenotypes}.}

\item{summarize}{Should the output be summarized?}
}
\value{
Summarized intensity measures with \code{group_var} column(s) and
\code{Phenotype} (named in \code{...}), \code{measure}, and \code{intensity}.
}
\description{
Select IF intensity measures to associate with a phneotype.
}
\seealso{
choose_measure
}
