% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_all_csd}
\alias{import_all_csd}
\title{Import All Cell Segmentation Data Files}
\usage{
import_all_csd(sample_dir, recursive = TRUE)
}
\arguments{
\item{sample_dir}{Directory or directories to search for inForm Cell Segmentation Files}

\item{recursive}{Should the directories be searched recursively?}
}
\value{
A single tibble containing all of the imported cell segmentation data
}
\description{
Import all cell segmentation files inside a folder or folders. Uses
\link[phenoptr:list_cell_seg_files]{phenoptr::list_cell_seg_files} and \link[phenoptr:read_cell_seg_data]{phenoptr::read_cell_seg_data} to import
the data sets, with additional but minor processing. All imported cell
segmentation files are row-bound upon import.
}
\section{Addtional Processing}{

\itemize{
\item Empty \code{Phenotype}s are relabelled as \code{"Undefined"}.
\item \code{Phenotype} values other than \code{"Undefined"} or \code{"Other"} are made uppercase for consistency.
\item \code{Phenotype} levels are converted to factors, wtih \code{"Other"} and \code{"Undefined"} last.
\item The file extension \code{.im3} is removed from the value in the \code{Sample Name} column.
\item The \code{Sample Name} column is renamed \code{Sample} for convenience.
}
}

\examples{
\dontrun{
import_all_csd("base_csd_dir")
}

}
