% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{merge_sample_info}
\alias{merge_sample_info}
\title{Merge Sample Information}
\usage{
merge_sample_info(csd, sample_info)
}
\arguments{
\item{csd}{Cell segmenation data from \code{\link[=import_all_csd]{import_all_csd()}}}

\item{sample_info}{A dataframe containing at least the column \code{Sample} and
additional columns containin information about each sample.}
}
\description{
Merge information about each sample into the CSD file. The sample info should
be a data frame containing at least the column \code{Sample}. Columns that
overlap with column names in the cell segmentation data are renamed with an
appended \code{.info}.
}
\examples{
\dontrun{
merge_sample_info(csd, sample_info)
}

}
