% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reorder_phenotype}
\alias{reorder_phenotype}
\title{Re-Order Phenotype}
\usage{
reorder_phenotype(x, sort = c("alphabetical", "frequency", "weighted"),
  ..., .var = Phenotype)
}
\arguments{
\item{x}{Input data frame}

\item{sort}{One of "alphabetical", "frequencuy", or "weighted". The latter
two use \link[forcats:fct_infreq]{forcats::fct_infreq} and \link[forcats:fct_reorder2]{forcats::fct_reorder2} respectively and
any additional parameters in \code{...} are passed to these functions.}

\item{.var}{Optional unquoted name for the Phenotype column, default is
\code{Phenotype}.}
}
\description{
Re-order the factor levels of the \code{.var} column (default is \code{Phenotype}) by
"alphabetical", "frequency", or "weighted". The "Other" and "Undefined"
levels are also automatically moved to the end of the ordering if present.
}
\examples{
\dontrun{
reorder_phenotype(csd, sort = "frequency")
}

}
