% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{unique_values}
\alias{unique_values}
\title{Get Unique Values of a Column}
\usage{
unique_values(x, variable, expected = NULL)
}
\arguments{
\item{x}{Data frame (grouped or ungrouped)}

\item{variable}{Column of interest}

\item{expected}{Expected values for \code{variable}}
}
\description{
Gets all unique column values, where the expected values for \code{variable} can
be declared in advance, while respecting grouping variables. Thus, if the
input is a grouped data frame, the output is all unique combinations of the
grouping columns crossed with the \code{variable} (or \code{expected}) values.
}
\seealso{
\link{count_observations}
}
